Converted int to Uint8 to resolve compiler warnings.

+++ source/core/font.cpp
@@ -58,7 +58,7 @@ bool font_class::load(std::string file_name, int pt_size)
     return(true);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text,int int_data)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text,int int_data)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -106,7 +106,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text,
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text,float float_data)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text,float float_data)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -154,7 +154,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text,
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -198,7 +198,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,std::string text)
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -246,7 +246,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
 
 
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,int int_data)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,int int_data)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -297,7 +297,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,float float_data)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,float float_data)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -348,7 +348,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,int int_data,std::string text_2)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,int int_data,std::string text_2)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -400,7 +400,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,float float_data,std::string text_2)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,float float_data,std::string text_2)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -452,7 +452,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,float float_data,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,float float_data,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -505,7 +505,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,unsigned long long ullint_data,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,unsigned long long ullint_data,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -558,7 +558,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,int int_data,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,int int_data,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -611,7 +611,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text_data,std::string text)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text_data,std::string text)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -660,7 +660,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,unsigned long long ullint_data)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,unsigned long long ullint_data)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -711,7 +711,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,unsigned long long ullint_data,std::string text_2)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,unsigned long long ullint_data,std::string text_2)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
@@ -763,7 +763,7 @@ bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs
     else return(false);
 };
 
-bool font_class::write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text,int int_data,std::string text_2,int int_data_2,std::string text_3)
+bool font_class::write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text,int int_data,std::string text_2,int int_data_2,std::string text_3)
 {
     if ((text.length() > 0) && (font_class::font_data != NULL))
     {
