+++ source/core/font.hpp
@@ -36,21 +36,21 @@ class font_class
        ~font_class(void);
         bool load(std::string file_name);
         bool load(std::string file_name, int pt_size);
-        bool write(int r,int g,int b,int a,float x,float y,std::string text,int int_data);
-        bool write(int r,int g,int b,int a,float x,float y,std::string text,float float_data);
-        bool write(int r,int g,int b,int a,float x,float y,std::string text);
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string text);
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,float              float_data  );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    ,std::string        text_2      );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,float              float_data  ,std::string        text_2      );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,float              float_data ,std::string        text        );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,unsigned long long ullint_data,std::string        text        );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,int                int_data   ,std::string        text        );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text_data  ,std::string        text        );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,unsigned long long ullint_data );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,unsigned long long ullint_data ,std::string        text_2      );
-        bool write(int r,int g,int b,int a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    ,std::string        text_2      ,int                int_data_2 ,std::string        text_3      );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text,int int_data);
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text,float float_data);
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,std::string text);
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string text);
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,float              float_data  );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    ,std::string        text_2      );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,float              float_data  ,std::string        text_2      );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,float              float_data ,std::string        text        );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,unsigned long long ullint_data,std::string        text        );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,int                int_data   ,std::string        text        );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text_data  ,std::string        text        );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,unsigned long long ullint_data );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,unsigned long long ullint_data ,std::string        text_2      );
+        bool write(Uint8 r,Uint8 g,Uint8 b,Uint8 a,float x,float y,float ws,float hs,std::string        text       ,int                int_data    ,std::string        text_2      ,int                int_data_2 ,std::string        text_3      );
 };
 
 #endif //FONT_H
